requirements.txt
flask
pyttsx3
SpeechRecognition
google-genai
